complete refactor of tests
