+++ src/hydrosar/hand/__init__.py
@@ -1,3 +1,4 @@
+from asf_tools.hydrosar.hand.calculate import calculate_hand_for_basins, make_copernicus_hand
 from asf_tools.hydrosar.hand.prepare import prepare_hand_vrt
 
 __all__ = [
